# storage/api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: LMS App API
  description: |
    API documentation for the Library Management System mobile/web app.

    **Authentication:** Use a Bearer Token (Laravel Sanctum) for protected routes.
    Obtain a token via `POST /api/v1/auth/login`.

    **Response format:** Most endpoints return `{ message, body, ... }` where `message` is `success` or `error`.
  version: 1.0.0
  contact:
    name: LMS Backend
servers:
  - url: /api
    description: Local server
...
paths:
...
  /books/search:
    get:
      tags: [Books]
      summary: Search books
      description: |
        Search and filter books. All query parameters are optional and can be combined.
        **Requires:** valid token.
      operationId: booksSearch
      security:
        - bearerAuth: []
      parameters:
        - name: ISBN
          in: query
          schema:
            type: integer
          description: Exact ISBN match
        - name: title
          in: query
          schema:
            type: string
          description: Partial title search (LIKE)
        - name: auther_id
          in: query
          schema:
            type: integer
          description: Filter by author ID
        - name: catagory_id
          in: query
          schema:
            type: integer
          description: Filter by category ID
        - name: publisher_id
          in: query
          schema:
            type: integer
          description: Filter by publisher ID
        - name: year
          in: query
          schema:
            type: string
          description: Filter by year of publishing
        - name: edition
          in: query
          schema:
            type: string
          description: Filter by edition number
      responses:
        '200':
          description: Books retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookSearchResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /books/{ISBN}:
    get:
      tags: [Books]
      summary: Get book by ISBN
      description: |
        Retrieve a single book with author, category, and publisher relations.
        **Requires:** valid token.
      operationId: booksShow
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/BookISBN'
      responses:
        '200':
          description: Book found
          content:
            application/json:
              schema:
                type: object
                properties:
                  book:
                    $ref: '#/components/schemas/Book'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Book not found
  /books/{ISBN}/copies:
    get:
      tags: [Books]
      summary: List book copies and their states
      description: |
        Returns all physical copies of a book with availability summary.
        **Requires:** valid token.
      operationId: booksCopies
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/BookISBN'
      responses:
        '200':
          description: Copies retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookCopiesResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Book not found
  /books/store:
    post:
      tags: [Books]
      summary: Add a new book
      description: |
        Create a book and the specified number of physical copies.
        **Requires:** valid token + LIBRARIAN or ADMIN role.
      operationId: booksStore
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/StoreBookRequest'
      responses:
        '201':
          description: Book created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreBookResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                type: object
                properties:
                  body:
                    type: string
                    example: Unauthorized
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LegacyErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LegacyErrorResponse'
  /books/update/{ISBN}:
    post:
      tags: [Books]
      summary: Update a book
      description: |
        Update book metadata. Only provided fields are updated.
        **Requires:** valid token + LIBRARIAN or ADMIN role.
      operationId: booksUpdate
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/BookISBN'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBookRequest'
      responses:
        '200':
          description: Book updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  body:
                    type: string
                    example: Book updated
                  book:
                    $ref: '#/components/schemas/Book'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                type: object
                properties:
                  body:
                    type: string
                    example: Unauthorized
        '404':
          description: Book not found
  /books/destroy/{ISBN}:
    post:
      tags: [Books]
      summary: Delete a book
      description: |
        Permanently delete a book record.
        **Requires:** valid token + ADMIN role.
      operationId: booksDestroy
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/BookISBN'
      responses:
        '200':
          description: Book deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  body:
                    type: string
                    example: Book deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Admin role required
          content:
            application/json:
              schema:
                type: object
                properties:
                  body:
                    type: string
                    example: Unauthorized
        '404':
          description: Book not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum Token
      description: Enter the token only (Swagger adds the Bearer prefix automatically)
  parameters:
...
    BookISBN:
      name: ISBN
      in: path
      required: true
      description: Book ISBN (primary key)
      schema:
        type: integer
        example: 9780141036144
  responses:
    Unauthorized:
      description: Unauthorized — missing or invalid token
      content:
        application/json:
          schema:
            oneOf:
              - type: object
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
              - $ref: '#/components/schemas/ApiErrorResponse'
  schemas:
...
    StoreBookRequest:
      type: object
      required:
        - ISBN
        - auther_id
        - catagory_id
        - publisher_id
        - title
        - discription
        - price
        - year_of_publishing
        - number_edition
        - copies_count
        - copy_condition
      properties:
        ISBN:
          type: integer
          description: Unique book ISBN
          example: 9780141036144
        auther_id:
          type: integer
          description: Author ID (authers table)
          example: 1
        catagory_id:
          type: integer
          description: Category ID (catagories table)
          example: 2
        publisher_id:
          type: integer
          description: Publisher ID
          example: 3
        title:
          type: string
          maxLength: 255
          example: "1984"
        discription:
          type: string
          example: "A dystopian social science fiction novel."
        price:
          type: number
          minimum: 0
          example: 25.50
        year_of_publishing:
          type: string
          maxLength: 4
          example: "1949"
        number_edition:
          type: string
          example: "1"
        copies_count:
          type: integer
          minimum: 1
          maximum: 500
          description: Number of physical copies to create
          example: 5
        copy_condition:
          type: string
          enum: [new, almost_new, worn]
          example: new
        cover_image:
          type: string
          format: binary
          description: Cover image (png, jpg, gif — max 4 MB)
    UpdateBookRequest:
      type: object
      properties:
        title:
          type: string
        discription:
          type: string
        price:
          type: number
          minimum: 0
        year_of_publishing:
          type: string
        number_edition:
          type: string
        auther_id:
          type: integer
        catagory_id:
          type: integer
        publisher_id:
          type: integer
...
    Author:
      type: object
      properties:
        id:
          type: integer
        firstname:
          type: string
        lastname:
          type: string
        nationality:
          type: string
          nullable: true
    Category:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        discription:
          type: string
          nullable: true
    Publisher:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        location:
          type: string
          nullable: true
    Book:
      type: object
      properties:
        ISBN:
          type: integer
        auther_id:
          type: integer
        catagory_id:
          type: integer
        publisher_id:
          type: integer
        title:
          type: string
        discription:
          type: string
        price:
          type: number
        amount:
          type: integer
          description: Total number of copies
        rate_avg:
          type: number
        cover_url:
          type: string
          nullable: true
        year_of_publishing:
          type: string
        number_edition:
          type: string
        author:
          $ref: '#/components/schemas/Author'
        category:
          $ref: '#/components/schemas/Category'
        publisher:
          $ref: '#/components/schemas/Publisher'
    BookCopy:
      type: object
      properties:
        instance_id:
          type: integer
        condition:
          type: string
          enum: [new, almost_new, worn]
        status:
          type: string
          description: Instance state (e.g. available, borrowed)
          example: available
...
    BookSearchResponse:
      type: object
      properties:
        total:
          type: integer
        books:
          type: array
          items:
            $ref: '#/components/schemas/Book'
    BookCopiesResponse:
      type: object
      properties:
        book:
          $ref: '#/components/schemas/Book'
        summary:
          type: object
          additionalProperties:
            type: integer
          description: Count of copies per status
          example:
            available: 3
            borrowed: 1
        copies:
          type: array
          items:
            $ref: '#/components/schemas/BookCopy'
    StoreBookResponse:
      type: object
      properties:
        body:
          type: string
          example: Book added successfully
        book:
          $ref: '#/components/schemas/Book'
        copies_created:
          type: integer
          example: 5
    ApiErrorResponse:
      type: object
      properties:
        message:
          type: string
          enum: [error]
        body:
          type: string
...
    LegacyErrorResponse:
      type: object
      properties:
        status:
          type: string
          enum: [error]
        body:
          type: string